refactor(kyosei): contextモジュールを責務ごとに分割
`context.ts`が型定義、GitHub URL解析、ローカルブランチ解決、ルーティングを全て担っており肥大化していました。
責務ごとにモジュールを分割し、循環importも解消しました。

- `context-type.ts`: 共通型定義(`PrIdentifier`, `GitHubOutputContext`, `LocalOutputContext`, `ReviewContext`)
- `context-github.ts`: PR URLの解析(`parsePrUrl`)
- `context-local.ts`: ローカル出力のブランチ解決(`resolveLocalContext`)
- `context.ts`: エントリポイント(`detectReviewContext`)のみ

テストも対応するモジュールごとに分割しました:
- `context-github.test.ts`: `parsePrUrl`の直接テスト
- `context-local.test.ts`: `resolveLocalContext`の直接テスト
- `context.test.ts`: `detectReviewContext`のルーティングテスト

利用側のimportは型を`context-type.ts`から、関数を各モジュールから直接importするように変更しました。
